move edit date and time button from SimulationPanel to ModulePanel. Change Temperature.java and DateTime.java to be accessible by class, not by object.
